Usa cidade fake no teste do controller

+++ test/controllers/util/WithAuthentication.java
@@ -27,9 +27,9 @@ public class WithAuthentication extends WithApplication {
     	admin = app.injector().instanceOf(JPAApi.class).withTransaction(()->{
     		Configuration configuration = app.injector().instanceOf(Configuration.class);
     		String adminEmail = configuration.getString(Cidadao.ADMIN_EMAIL);
-    		Cidadao cidadao = new Cidadao("Governo Federal", adminEmail);
     		CidadaoDAO cidadaoDAO = app.injector().instanceOf(CidadaoDAO.class);
     		if(cidadaoDAO.findByLogin(adminEmail) == null){
+    			Cidadao cidadao = new Cidadao("Governo Federal", adminEmail);
 				return cidadaoDAO.saveAndUpdate(cidadao);
     		}
     		return cidadaoDAO.findByLogin(adminEmail);
@@ -46,7 +46,9 @@ public class WithAuthentication extends WithApplication {
     		String adminEmail = configuration.getString(Cidadao.ADMIN_EMAIL);
     		CidadaoDAO cidadaoDAO = app.injector().instanceOf(CidadaoDAO.class);
     		Cidadao admin = cidadaoDAO.findByLogin(adminEmail);
-    		cidadaoDAO.remove(admin);
+    		if(admin != null){
+    			cidadaoDAO.remove(admin);
+    		}
 //    		cidadaoDAO.saveAndUpdate(admin);
     	});
     }
